fix: 修録音の量子化ビット深度とチャンネル数の型を修正

+++ VoiLog/Recording/RecordingFeature.swift
@@ -112,8 +112,8 @@ struct RecordingFeature {
             title: title,
             fileFormat: fileFormat,
             samplingFrequency: UserDefaultsManager.shared.samplingFrequency,
-            quantizationBitDepth: Int16(UserDefaultsManager.shared.quantizationBitDepth),
-            numberOfChannels: Int16(UserDefaultsManager.shared.numberOfChannels),
+            quantizationBitDepth: UserDefaultsManager.shared.quantizationBitDepth,
+            numberOfChannels: UserDefaultsManager.shared.numberOfChannels,
             url: recordingUrl
           )
           voiceMemoRepository.insert(recordingVoice)
@@ -143,8 +143,8 @@ struct RecordingFeature {
             title: title,
             fileFormat: fileFormat,
             samplingFrequency: UserDefaultsManager.shared.samplingFrequency,
-            quantizationBitDepth: Int16(UserDefaultsManager.shared.quantizationBitDepth),
-            numberOfChannels: Int16(UserDefaultsManager.shared.numberOfChannels),
+            quantizationBitDepth: UserDefaultsManager.shared.quantizationBitDepth,
+            numberOfChannels: UserDefaultsManager.shared.numberOfChannels,
             url: recordingUrl
           )
           voiceMemoRepository.insert(recordingVoice)
